Rename zh to zh-cn; Fix localize mixin files
