+++ components/d2l-activity-card/lang/zh-cn.js
