+++ components/d2l-activity-date/lang/zh-cn.js
