+++ components/d2l-activity-editor/lang/zh-cn.js
