+++ components/d2l-activity-evaluation-icon/ActivityEvaluationIconBaseLocalize.js
@@ -1,12 +1,11 @@
 import 'd2l-localize-behavior/d2l-localize-behavior.js';
 import ar from './lang/ar.js';
-import cygb from './lang/cy-gb.js';
-import dadk from './lang/da-dk.js';
+import cy from './lang/cy.js';
+import da from './lang/da.js';
 import de from './lang/de.js';
 import en from './lang/en.js';
 import es from './lang/es.js';
 import eses from './lang/es-es.js';
-import fi from './lang/fi.js';
 import fr from './lang/fr.js';
 import frfr from './lang/fr-fr.js';
 import ja from './lang/ja.js';
@@ -15,7 +14,7 @@ import nl from './lang/nl.js';
 import pt from './lang/pt.js';
 import sv from './lang/sv.js';
 import tr from './lang/tr.js';
-import zh from './lang/zh.js';
+import zhcn from './lang/zh-cn.js';
 import zhtw from './lang/zh-tw.js';
 
 // eslint-disable-next-line sort-imports
@@ -32,12 +31,11 @@ const ActivityEvaluationIconBaseLocalizeImpl = (superClass) => {
 						return {
 							'en': en,
 							'ar': ar,
-							'cy-gb': cygb,
-							'da-dk': dadk,
+							'cy': cy,
+							'da': da,
 							'de': de,
 							'es': es,
 							'es-es': eses,
-							'fi': fi,
 							'fr': fr,
 							'fr-fr': frfr,
 							'ja': ja,
@@ -46,7 +44,7 @@ const ActivityEvaluationIconBaseLocalizeImpl = (superClass) => {
 							'pt': pt,
 							'sv': sv,
 							'tr': tr,
-							'zh': zh,
+							'zh-cn': zhcn,
 							'zh-tw': zhtw
 						};
 					}
