+++ components/d2l-activity-evaluation-icon/lang/zh-cn.js
