+++ components/d2l-activity-list-item/lang/zh-cn.js
