+++ components/d2l-quick-eval-widget/lang/zh-cn.js
