+++ components/d2l-quick-eval/QuickEvalLocalize.js
@@ -1,12 +1,11 @@
 import 'd2l-localize-behavior/d2l-localize-behavior.js';
 import ar from './lang/ar.js';
-import cygb from './lang/cy-gb.js';
-import dadk from './lang/da-dk.js';
+import cy from './lang/cy.js';
+import da from './lang/da.js';
 import de from './lang/de.js';
 import en from './lang/en.js';
 import es from './lang/es.js';
 import eses from './lang/es-es.js';
-import fi from './lang/fi.js';
 import fr from './lang/fr.js';
 import frfr from './lang/fr-fr.js';
 import ja from './lang/ja.js';
@@ -15,7 +14,7 @@ import nl from './lang/nl.js';
 import pt from './lang/pt.js';
 import sv from './lang/sv.js';
 import tr from './lang/tr.js';
-import zh from './lang/zh.js';
+import zhcn from './lang/zh-cn.js';
 import zhtw from './lang/zh-tw.js';
 
 // eslint-disable-next-line sort-imports
@@ -33,12 +32,11 @@ const QuickEvalLocalizeImpl = (superClass) => {
 						return {
 							'en': en,
 							'ar': ar,
-							'cy-gb': cygb,
-							'da-dk': dadk,
+							'cy': cygb,
+							'da': dadk,
 							'de': de,
 							'es': es,
 							'es-es': eses,
-							'fi': fi,
 							'fr': fr,
 							'fr-fr': frfr,
 							'ja': ja,
@@ -47,7 +45,7 @@ const QuickEvalLocalizeImpl = (superClass) => {
 							'pt': pt,
 							'sv': sv,
 							'tr': tr,
-							'zh': zh,
+							'zh-cn': zhcn,
 							'zh-tw': zhtw
 						};
 					}
@@ -76,11 +74,11 @@ export const LitQuickEvalLocalize = superclass => class extends LocalizeMixin(su
 				case 'ar':
 					translations = ar;
 					break;
-				case 'cy-gb':
-					translations = cygb;
+				case 'cy':
+					translations = cy;
 					break;
-				case 'da-dk':
-					translations = dadk;
+				case 'da':
+					translations = da;
 					break;
 				case 'de':
 					translations = de;
@@ -121,8 +119,8 @@ export const LitQuickEvalLocalize = superclass => class extends LocalizeMixin(su
 				case 'tr':
 					translations = tr;
 					break;
-				case 'zh':
-					translations = zh;
+				case 'zh-cn':
+					translations = zhcn;
 					break;
 				case 'zh-tw':
 					translations = zhtw;
